合并

+++ cocos/editor-support/spine/SkeletonBatch.cpp
@@ -146,8 +146,8 @@ void SkeletonBatch::reset() {
 
 cocos2d::TrianglesCommand* SkeletonBatch::nextFreeCommand() {
 	if (_commandsPool.size() <= _nextFreeCommand) {
-		unsigned int newSize = _commandsPool.size() * 2 + 1;
-		for (int i = _commandsPool.size();  i < newSize; i++) {
+		size_t newSize = _commandsPool.size() * 2 + 1;
+		for (size_t i = _commandsPool.size();  i < newSize; i++) {
 			_commandsPool.push_back(new TrianglesCommand());
 		}
 	}
